+++ cmsat/sccfinder.h
@@ -102,7 +102,7 @@ class SCCFinder {
         };
 
         const Stats& getStats() const;
-        uint64_t memUsed() const;
+        size_t memUsed() const;
 
     private:
