+++ cmsat/xorfinder.h
@@ -277,7 +277,7 @@ public:
 
     const Stats& getStats() const;
     size_t getNumCalls() const;
-    virtual uint64_t memUsed() const;
+    virtual size_t memUsed() const;
 
 private:
